+++ ldlidar_component/component/src/ldlidar_component.cpp
@@ -1,4 +1,4 @@
-//  Copyright 2022 Walter Lucetti
+//  Copyright 2024 Walter Lucetti
 //
 //  Licensed under the Apache License, Version 2.0 (the "License");
 //  you may not use this file except in compliance with the License.