+++ ldlidar_node/launch/ldlidar_bringup.launch.py
@@ -19,11 +19,16 @@ from ament_index_python.packages import get_package_share_directory
 from launch import LaunchDescription
 from launch.actions import DeclareLaunchArgument
 from launch.substitutions import LaunchConfiguration
-from launch_ros.actions import Node, LifecycleNode
+from launch_ros.actions import (
+    Node,
+    ComposableNodeContainer
+)
+from launch_ros.descriptions import ComposableNode
 
 
 def generate_launch_description():
-    
+
+    node_ns = LaunchConfiguration('node_namespace')
     node_name = LaunchConfiguration('node_name')
 
     # Lidar node configuration file
@@ -34,23 +39,47 @@ def generate_launch_description():
     )
 
     # Launch arguments
+    declare_node_namespace_cmd = DeclareLaunchArgument(
+        'node_namespace',
+        default_value='',
+        description='Namespace of the node'
+    )
     declare_node_name_cmd = DeclareLaunchArgument(
         'node_name',
         default_value='ldlidar_node',
         description='Name of the node'
     )
 
-    # LDLidar lifecycle node
-    ldlidar_node = LifecycleNode(
-        package='ldlidar_node',
-        executable='ldlidar_node',
-        name=node_name,
-        namespace='',
-        output='screen',
-        parameters=[
-            # YAML files
-            lidar_config_path  # Parameters
-        ]
+    # LDLidar component
+    ldlidar_component = ComposableNode(
+            package='ldlidar_component',
+            namespace=node_ns,
+            plugin='ldlidar::LdLidarComponent',
+            name=node_name,
+            parameters=[
+                # YAML files
+                lidar_config_path  # Parameters
+            ],
+            extra_arguments=[{'use_intra_process_comms': True}]
+        )
+
+    # ROS 2 Component Container
+    distro = os.environ['ROS_DISTRO']
+    if distro == 'foxy':
+        # Foxy does not support the isolated mode
+        container_exec='component_container'
+    else:
+        container_exec='component_container_isolated'
+    
+    ldlidar_container = ComposableNodeContainer(
+            name='zed_container',
+            namespace=node_ns,
+            package='rclcpp_components',
+            executable=container_exec,
+            composable_node_descriptions=[
+                ldlidar_component
+            ],
+            output='screen',
     )
 
     # URDF path
@@ -66,6 +95,7 @@ def generate_launch_description():
     rsp_node = Node(
         package='robot_state_publisher',
         executable='robot_state_publisher',
+        namespace=node_ns,
         name='ldlidar_state_publisher',
         output='screen',
         parameters=[{'robot_description': robot_desc}],
@@ -76,12 +106,13 @@ def generate_launch_description():
     ld = LaunchDescription()
 
     # Launch arguments
+    ld.add_action(declare_node_namespace_cmd)
     ld.add_action(declare_node_name_cmd)
 
-    # Launch Nav2 Lifecycle Manager
+    # Launch Robot State Publisher node
     ld.add_action(rsp_node)
 
-    # LDLidar Lifecycle node
-    ld.add_action(ldlidar_node)
+    # LDLidar Lifecycle node in container
+    ld.add_action(ldlidar_container)
 
     return ld